Update dummy_creation.py
Make the function take in two additional arguments - the name of the salesperson (string), and a date the named salesperson joined the company (in the "yyyy-mm-dd 00:00:00" format). Also, make the "timestamp" data to be generated as a random date between the date the salesperson joined the company and a random point later in time.